+++ services/QuillLMS/app/assets/stylesheets/pages/premium_hub/usage-snapshot-report.scss
@@ -51,6 +51,7 @@
     .filter-label {
       font-size: 16px;
       line-height: 22px;
+      margin-bottom: 16px;
       &:not(:first-of-type) {
         margin-top: 24px;
       }
@@ -76,6 +77,13 @@
         margin-bottom: 16px;
       }
     }
+    .dropdown-container {
+      min-height: 46px;;
+      justify-content: flex-start;
+    }
+    .dropdown__menu {
+      width: 258px;
+    }
   }
   main {
     padding: 32px 64px;
@@ -91,7 +99,7 @@
       font-size: 24px;
     }
     .tabs {
-      margin-top: 32px;
+      margin-top: 24px;
       display: flex;
       .tab {
         padding: 0px 32px;
@@ -140,7 +148,7 @@
     .snapshot-section {
       box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.15);
       border-radius: 8px;
-      margin-top: 32px;
+      margin-top: 24px;
       padding: 24px;
       h2 {
         font-weight: 600;
@@ -248,6 +256,9 @@
           font-size: 14px;
           line-height: 20px;
           color: #373737;
+        }
+        &:hover, &:focus {
+          text-decoration: none;
         }
       }
       .snapshot-ranking {
