+++ services/QuillLMS/app/views/pages/shared/_footer.html.erb
@@ -1,7 +1,7 @@
 <% footer_class = current_user ? 'home-footer white' : 'home-footer' %>
 <% logo_class = current_user ? 'dark' : 'light' %>
 
-<% if show_site_navigation? %>
+<% if show_site_navigation? && request.path.exclude?('usage_snapshot') %>
   <% if current_user&.student? || current_user&.staff? %>
     <footer class="student-footer">
       <div class="inner-container">
